move auth config to separate file
